Move certificate logic to NodeRuntime
